backend/requirements.txt
fastapi
uvicorn
openai
pinecone
langchain
langchain-community
tiktoken
python-dotenv
